improve pack sh

+++ ArduCopter/pack_vers.sh
@@ -1,7 +1,19 @@
 # !bin/bash
+echo "========================================="
 date
+
+if [ $# = 0 ]
+then
+    echo "Usage: $1 <FW-version>"
+    echo "\n=========================================\n"
+    exit 
+fi
+
+export FW_NAME=FW-V${1}-`date +%Y%m%d`.fc
+echo "package ${FW_NAME} start\n"
+
 rm -rf APP-new
-rm -f FW-v0.2* 
+rm -f FW-v* 
 mkdir -p APP-new
 make pxf
 cp ArduCopter.elf APP-new/app_quad
@@ -10,6 +22,5 @@ cp ArduCopter.elf APP-new/app_hexa
 make pxf-octa
 cp ArduCopter.elf APP-new/app_octa
 cp readme APP-new/
-export FW_NAME=FW-v0.2-`date +%Y%m%d`.fc
 tar -cf ${FW_NAME} APP-new
 echo "package FW ${FW_NAME} done\n========================================="
